--- x/outline-connectivity/access_key.go
@@ -1,94 +1,0 @@
-// Copyright 2023 Jigsaw Operations LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import (
-	"encoding/base64"
-	"fmt"
-	"net"
-	"net/url"
-	"strconv"
-	"strings"
-
-	"github.com/Jigsaw-Code/outline-sdk/transport/shadowsocks"
-)
-
-type sessionConfig struct {
-	Hostname  string
-	Port      int
-	CryptoKey *shadowsocks.EncryptionKey
-	Prefix    Prefix
-}
-
-type Prefix []byte
-
-func (p Prefix) String() string {
-	runes := make([]rune, len(p))
-	for i, b := range p {
-		runes[i] = rune(b)
-	}
-	return string(runes)
-}
-
-// TODO(fortuna): provide this as a reusable library. Perhaps as x/shadowsocks or x/outline.
-func parseAccessKey(accessKey string) (*sessionConfig, error) {
-	var config sessionConfig
-	accessKeyURL, err := url.Parse(accessKey)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse access key: %w", err)
-	}
-	var portString string
-	// Host is a <host>:<port> string
-	config.Hostname, portString, err = net.SplitHostPort(accessKeyURL.Host)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse endpoint address: %w", err)
-	}
-	config.Port, err = strconv.Atoi(portString)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse port number: %w", err)
-	}
-	cipherInfoBytes, err := base64.URLEncoding.WithPadding(base64.NoPadding).DecodeString(accessKeyURL.User.String())
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode cipher info [%v]: %v", accessKeyURL.User.String(), err)
-	}
-	cipherName, secret, found := strings.Cut(string(cipherInfoBytes), ":")
-	if !found {
-		return nil, fmt.Errorf("invalid cipher info: no ':' separator")
-	}
-	config.CryptoKey, err = shadowsocks.NewEncryptionKey(cipherName, secret)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create cipher: %w", err)
-	}
-	prefixStr := accessKeyURL.Query().Get("prefix")
-	if len(prefixStr) > 0 {
-		config.Prefix, err = ParseStringPrefix(prefixStr)
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse prefix: %w", err)
-		}
-	}
-	return &config, nil
-}
-
-func ParseStringPrefix(utf8Str string) (Prefix, error) {
-	runes := []rune(utf8Str)
-	rawBytes := make([]byte, len(runes))
-	for i, r := range runes {
-		if (r & 0xFF) != r {
-			return nil, fmt.Errorf("character out of range: %d", r)
-		}
-		rawBytes[i] = byte(r)
-	}
-	return rawBytes, nil
-}
